groundbeef {fitdistrplus}	R Documentation

Ground beef serving size data set

Description

Serving sizes collected in a French survey, for ground beef patties consumed by children under 5 years old.

Usage

data(groundbeef)

Format

groundbeef is a data frame with 1 column (serving: serving sizes in grams)

Source

Delignette-Muller, M.L., Cornu, M. 2008. Quantitative risk assessment for Escherichia coli O157:H7 in frozen ground beef patties consumed by young children in French households. International Journal of Food Microbiology, 128, 158-164.

Examples

# (1) load of data
#
data(groundbeef)

# (2) description and plot of data
#
serving <- groundbeef$serving
descdist(serving)
plotdist(serving)

# (3) fit of a Weibull distribution to data
#
fitW <- fitdist(serving,"weibull")
summary(fitW)
plot(fitW)
gofstat(fitW)

[Package fitdistrplus version 1.0-14 Index]