smokedfish {fitdistrplus}	R Documentation

Contamination data of Listeria monocytogenes in smoked fish

Description

Contamination data of Listeria monocytogenes in smoked fish on the Belgian market in the period 2005 to 2007.

Usage

data(smokedfish)

Format

smokedfish is a data frame with 2 columns named left and right, describing each observed value of Listeria monocytogenes concentration (in CFU/g) as an interval. The left column contains either NA for left censored observations, the left bound of the interval for interval censored observations, or the observed value for non-censored observations. The right column contains either NA for right censored observations, the right bound of the interval for interval censored observations, or the observed value for non-censored observations.

Source

Busschaert, P., Geereard, A.H., Uyttendaele, M., Van Impe, J.F., 2010. Estimating distributions out of qualitative and (semi) quantitative microbiological contamination data for use in risk assessment. International Journal of Food Microbiology. 138, 260-269.

Examples

# (1) load of data
#
data(smokedfish)

# (2) plot of data in CFU/g
#
plotdistcens(smokedfish)

# (3) plot of transformed data in log10[CFU/g]
#
Clog10 <- log10(smokedfish)
plotdistcens(Clog10)

# (4) Fit of a normal distribution to data in log10[CFU/g]
#
fitlog10 <- fitdistcens(Clog10,"norm")
summary(fitlog10)
plot(fitlog10)

[Package fitdistrplus version 1.0-14 Index]